# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TourFlow Real Estate Tour Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours:
    get:
      tags: [tours]
      operationId: listTours
      summary: List all tours for the current agent
      responses:
        '200':
          description: List of tours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourListResponse'
    post:
      tags: [tours]
      operationId: createTour
      summary: Create a new tour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTourRequest'
      responses:
        '201':
          description: Tour created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
  /tours/{tourId}:
    get:
      tags: [tours]
      operationId: getTour
      summary: Get a tour with all its stops
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      responses:
        '200':
          description: Tour found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourDetailResponse'
        '404':
          description: Tour not found
    put:
      tags: [tours]
      operationId: updateTour
      summary: Update a tour
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTourRequest'
      responses:
        '200':
          description: Tour updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
    delete:
      tags: [tours]
      operationId: deleteTour
      summary: Delete a tour
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      responses:
        '204':
          description: Tour deleted
  /tours/{tourId}/properties:
    post:
      tags: [tours]
      operationId: addPropertyToTour
      summary: Add a property to a tour (creates tour stop)
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPropertyToTourRequest'
      responses:
        '201':
          description: Property added to tour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourStopResponse'
  /tours/{tourId}/optimize:
    post:
      tags: [tours]
      operationId: optimizeTourRoute
      summary: Optimize the route order for a tour
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeRouteRequest'
      responses:
        '200':
          description: Route optimized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizeRouteResponse'
  /tours/{tourId}/stops/order:
    put:
      tags: [tours]
      operationId: reorderTourStops
      summary: Save a manually reordered stop sequence
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderStopsRequest'
      responses:
        '200':
          description: Order saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /tours/{tourId}/skip-stop:
    post:
      tags: [tours]
      operationId: skipTourStop
      summary: Skip a stop and reroute to the next best stop
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkipStopRequest'
      responses:
        '200':
          description: Stop skipped, next stop returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkipStopResponse'
  /tours/{tourId}/publish:
    post:
      tags: [tours]
      operationId: publishTour
      summary: Publish a tour to mobile
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      responses:
        '200':
          description: Tour published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourResponse'
  /tours/{tourId}/readiness:
    get:
      tags: [tours]
      operationId: getTourReadiness
      summary: Get the readiness summary for a tour
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      responses:
        '200':
          description: Tour readiness summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourReadinessResponse'
  /tours/{tourId}/generate-summary:
    post:
      tags: [tours]
      operationId: generateTourSummary
      summary: Generate an AI end-of-tour summary
      parameters:
        - $ref: '#/components/parameters/TourIdParam'
      responses:
        '200':
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TourSummaryResponse'
...
  /voice-notes/upload:
    post:
      tags: [voiceNotes]
      operationId: uploadVoiceNote
      summary: Upload a voice recording (multipart/form-data)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [tourStopId, audio]
              properties:
                tourStopId:
                  type: string
                audio:
                  type: string
                  format: binary
                durationSeconds:
                  type: integer
      responses:
        '201':
          description: Voice note uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceNoteResponse'
  /voice-notes/{voiceNoteId}/transcribe:
    post:
      tags: [voiceNotes]
      operationId: transcribeVoiceNote
      summary: Trigger transcription for a voice note
      parameters:
        - $ref: '#/components/parameters/VoiceNoteIdParam'
      responses:
        '200':
          description: Transcription job started or completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceNoteResponse'
  /voice-notes/{voiceNoteId}:
    get:
      tags: [voiceNotes]
      operationId: getVoiceNote
      summary: Get a voice note and its transcript
      parameters:
        - $ref: '#/components/parameters/VoiceNoteIdParam'
      responses:
        '200':
          description: Voice note details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceNoteDetailResponse'

  # ── Admin AI Config ─────────────────────────────────────────────────────────
...
components:
  parameters:
...
    TourIdParam:
      name: tourId
      in: path
      required: true
      schema:
        type: string
...
    VoiceNoteIdParam:
      name: voiceNoteId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── Auth schemas ──────────────────────────────────────────────────────────
...
    Buyer:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Tour:
      type: object
      required: [id, agentId, title, date, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        buyerId:
          type: [string, 'null']
        title:
          type: string
        date:
          type: string
        startTime:
          type: [string, 'null']
        startAddress:
          type: [string, 'null']
        startLat:
          type: [number, 'null']
        startLng:
          type: [number, 'null']
        endAddress:
          type: [string, 'null']
        endLat:
          type: [number, 'null']
        endLng:
          type: [number, 'null']
        buyerNotes:
          type: [string, 'null']
        geographicArea:
          type: [string, 'null']
        tags:
          type: [array, 'null']
          items:
            type: string
        status:
          type: string
          enum: [draft, active, published, completed, cancelled]
        publishedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TourResponse:
      type: object
      required: [tour]
      properties:
        tour:
          $ref: '#/components/schemas/Tour'
    TourListResponse:
      type: object
      required: [tours]
      properties:
        tours:
          type: array
          items:
            $ref: '#/components/schemas/Tour'
    CreateTourRequest:
      type: object
      required: [title, date]
      properties:
        buyerId:
          type: string
        title:
          type: string
        date:
          type: string
        startTime:
          type: string
        startAddress:
          type: string
        startLat:
          type: number
        startLng:
          type: number
        endAddress:
          type: string
        endLat:
          type: number
        endLng:
          type: number
        buyerNotes:
          type: string
        geographicArea:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateTourRequest:
      type: object
      properties:
        buyerId:
          type: string
        title:
          type: string
        date:
          type: string
        startTime:
          type: string
        startAddress:
          type: string
        startLat:
          type: number
        startLng:
          type: number
        endAddress:
          type: string
        endLat:
          type: number
        endLng:
          type: number
        buyerNotes:
          type: string
        geographicArea:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, published, completed, cancelled]

    # ── TourStop schemas ──────────────────────────────────────────────────────
...
    TourStop:
      type: object
      required: [id, tourId, propertyId, sequence, approvedStatus, skipped, visited, followUpFlag, revisitFlag, createdAt, updatedAt]
      properties:
        id:
          type: string
        tourId:
          type: string
        propertyId:
          type: string
        sequence:
          type: integer
        approvedStatus:
          type: string
          enum: [not_requested, requested, pending, approved, declined, needs_follow_up, restricted, cancelled]
        skipped:
          type: boolean
        skipReason:
          type: [string, 'null']
        skipNotes:
          type: [string, 'null']
        visited:
          type: boolean
        arrivalTime:
          type: [string, 'null']
          format: date-time
        departureTime:
          type: [string, 'null']
          format: date-time
        buyerInterest:
          type: [integer, 'null']
        kitchenRating:
          type: [integer, 'null']
        primarySuiteRating:
          type: [integer, 'null']
        backyardRating:
          type: [integer, 'null']
        roadNoiseRating:
          type: [integer, 'null']
        overallFitRating:
          type: [integer, 'null']
        followUpFlag:
          type: boolean
        revisitFlag:
          type: boolean
        quickTags:
          type: [array, 'null']
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TourStopResponse:
      type: object
      required: [stop]
      properties:
        stop:
          $ref: '#/components/schemas/TourStop'
...
    TourDetailResponse:
      type: object
      required: [tour, stops]
      properties:
        tour:
          $ref: '#/components/schemas/Tour'
        stops:
          type: array
          items:
            $ref: '#/components/schemas/TourStop'
        buyer:
          oneOf:
            - $ref: '#/components/schemas/Buyer'
            - type: 'null'
    AddPropertyToTourRequest:
      type: object
      required: [formattedAddress]
      properties:
        formattedAddress:
          type: string
        placeId:
          type: string
        lat:
          type: number
        lng:
          type: number
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        mlsId:
          type: string
        listPrice:
          type: integer
        beds:
          type: integer
        baths:
          type: number
        squareFeet:
          type: integer
        nickname:
          type: string
        notes:
          type: string
        propertyId:
          type: string
          description: Reuse an existing property by ID (skips property creation)
...
    OptimizeRouteRequest:
      type: object
      properties:
        includeApprovedOnly:
          type: boolean
          default: true
    OptimizeRouteResponse:
      type: object
      required: [orderedStopIds, estimatedDriveTimeMinutes]
      properties:
        orderedStopIds:
          type: array
          items:
            type: string
        estimatedDriveTimeMinutes:
          type: integer
    ReorderStopsRequest:
      type: object
      required: [orderedStopIds]
      properties:
        orderedStopIds:
          type: array
          items:
            type: string
    SkipStopRequest:
      type: object
      required: [stopId, reason]
      properties:
        stopId:
          type: string
        reason:
          type: string
          enum: [not_approved, client_changed_mind, running_late, access_issue, traffic, duplicate_choice, other]
        notes:
          type: string
    SkipStopResponse:
      type: object
      required: [skippedStop]
      properties:
        skippedStop:
          $ref: '#/components/schemas/TourStop'
        nextStop:
          oneOf:
            - $ref: '#/components/schemas/TourStop'
            - type: 'null'
    TourReadinessResponse:
      type: object
      required: [approvedCount, pendingCount, declinedCount, missingAgentInfoCount, estimatedDriveTimeMinutes, estimatedTotalMinutes]
      properties:
        approvedCount:
          type: integer
        pendingCount:
          type: integer
        declinedCount:
          type: integer
        missingAgentInfoCount:
          type: integer
        unresolved_restrictions:
          type: integer
        estimatedDriveTimeMinutes:
          type: integer
        estimatedTotalMinutes:
          type: integer

    # ── Showing Request schemas ───────────────────────────────────────────────
...
    VoiceNote:
      type: object
      required: [id, tourStopId, fileUrl, transcriptionStatus, createdAt, updatedAt]
      properties:
        id:
          type: string
        tourStopId:
          type: string
        fileUrl:
          type: string
        durationSeconds:
          type: [integer, 'null']
        transcriptionStatus:
          type: string
          enum: [pending, in_progress, completed, failed]
        typedNote:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VoiceNoteResponse:
      type: object
      required: [voiceNote]
      properties:
        voiceNote:
          $ref: '#/components/schemas/VoiceNote'
    VoiceNoteDetailResponse:
      type: object
      required: [voiceNote]
      properties:
        voiceNote:
          $ref: '#/components/schemas/VoiceNote'
        transcript:
          oneOf:
            - $ref: '#/components/schemas/Transcript'
            - type: 'null'
    Transcript:
      type: object
      required: [id, voiceNoteId, text, createdAt]
      properties:
        id:
          type: string
        voiceNoteId:
          type: string
        text:
          type: string
        provider:
          type: [string, 'null']
        confidence:
          type: [number, 'null']
        createdAt:
          type: string
          format: date-time

    # ── Summary schemas ───────────────────────────────────────────────────────
...
    TourSummary:
      type: object
      required: [id, tourId, summaryText, generatedAt, createdAt]
      properties:
        id:
          type: string
        tourId:
          type: string
        summaryText:
          type: string
        topHomes:
          type: [array, 'null']
          items:
            type: string
        homesToEliminate:
          type: [array, 'null']
          items:
            type: string
        buyerPreferences:
          type: [array, 'null']
          items:
            type: string
        nextActions:
          type: [array, 'null']
          items:
            type: string
        provider:
          type: [string, 'null']
        generatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TourSummaryResponse:
      type: object
      required: [summary]
      properties:
        summary:
          $ref: '#/components/schemas/TourSummary'

    # ── AI Config schemas ─────────────────────────────────────────────────────
...