# lib/api-zod/src/generated/types/optimizeRouteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TourFlow Real Estate Tour Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface OptimizeRouteRequest {
  includeApprovedOnly?: boolean;
}